We can now receive interrupts

+++ types.h
@@ -5,7 +5,7 @@ typedef char int8_t;
 typedef unsigned char uint8_t;
 typedef short int16_t;
 typedef unsigned short uint16_t;
-typedef long long int int32_t;
-typedef unsigned long long int uint32_t;
+typedef long int int32_t;
+typedef unsigned long int uint32_t;
 
 #endif
